+++ src/main/kotlin/com/github/hechtcarmel/jetbrainsindexmcpplugin/handlers/OptimizedSymbolSearch.kt
@@ -1,5 +1,6 @@
 package com.github.hechtcarmel.jetbrainsindexmcpplugin.handlers
 
+import com.github.hechtcarmel.jetbrainsindexmcpplugin.util.ProjectUtils
 import com.intellij.navigation.ChooseByNameContributor
 import com.intellij.navigation.ChooseByNameContributorEx
 import com.intellij.navigation.NavigationItem
@@ -213,8 +214,7 @@ object OptimizedSymbolSearch {
         }
 
         val file = targetElement.containingFile?.virtualFile ?: return null
-        val basePath = project.basePath ?: ""
-        val relativePath = file.path.removePrefix(basePath).removePrefix("/")
+        val relativePath = ProjectUtils.getToolFilePath(project, file)
 
         if (isExcludedPath(relativePath)) return null